api_service/pyproject.toml
[project]
name = "api-service"
version = "0.1.0"
description = "API Service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "opentelemetry-distro>=0.58b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.37.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
]